+++ src/app/services/authentication-guard.service.ts
@@ -1,0 +1,28 @@
+import { Injectable } from '@angular/core';
+import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
+import { AuthenticationService } from './authentication.service';
+
+
+@Injectable({
+  providedIn: 'root'
+})
+
+export class AuthenticationGuardService implements CanActivate {
+
+  constructor(private authService: AuthenticationService, private router: Router) { }
+
+  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    return this.authService.isLoggedIn.pipe(
+      map(e => {
+        if (e) {
+          return true;
+        } else {
+          this.router.navigate(['/login']);
+          return false;
+        }
+      })
+    );
+  }
+}